functions/requirements.txt
firebase_functions~=0.1.0
openai
pydantic
numpy
scikit-learn
umap-learn